# Specifiche/ServerForge_OpenAPI_Spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Server Forge",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://serverforge.cx/api"
    }
  ],
  "security": [
    {
      "http": []
    }
  ],
  "paths": {
...
    "/machines/{machine}": {
      "get": {
        "operationId": "machines.show",
        "summary": "Get a specific machine with detailed information",
        "tags": ["UserResource"],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "description": "The machine ID",
            "schema": { "type": "integer" }
          }
        ]
      }
    },
    "/containers/{container}": {
      "get": {
        "operationId": "containers.show",
        "summary": "Get a specific container with detailed information",
        "tags": ["UserResource"],
        "parameters": [
          {
            "name": "container",
            "in": "path",
            "required": true,
            "description": "The container ID",
            "schema": { "type": "integer" }
          }
        ]
      }
    },
    "/clusters/{cluster}": {
      "get": {
        "operationId": "clusters.show",
        "summary": "Get a specific cluster with detailed information",
        "tags": ["UserResource"],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "description": "The cluster ID",
            "schema": { "type": "integer" }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "http": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}